API Documentation

Below is the documentation for all functions exported by Onda.jl. For general information regarding the Onda format, please see beacon-biosignals/OndaFormat.

Note that Onda.jl's API follows a specific philosophy with respect to property access: users are generally expected to access fields via Julia's object.fieldname syntax, but should only mutate objects via the exposed API methods documented below.

Dataset

Onda.Dataset — Type

Dataset(path, custom_type=Any; create=false, strict=())

Return a Dataset instance that contains all metadata necessary to read and write to the Onda dataset stored at path. Note that this constuctor loads all the Recording objects contained in path/recordings.msgpack.zst.

custom_type is the typeof of the custom value found in each Recording object in the dataset.

If create is true, then an empty Onda dataset will be created at path.

The strict keyword argument is forwarded to MsgPack.unpack when that function is called while parsing path/recordings.msgpack.zst (see the MsgPack documentation for details regarding strict).

Onda.create_recording! — Function

create_recording!(dataset::Dataset{C}, duration::Nanosecond, custom=nothing)

Create uuid::UUID => recording::Recording where recording is constructed via the provided duration and custom fields, add the pair to dataset.recordings, and return the pair.

The custom argument is passed along to the Recording{C} constructor, such that custom isa C must hold true.

Onda.load — Function

load(dataset::Dataset, uuid::UUID, name::Symbol[, span::AbstractTimeSpan])

Load and return the Samples object corresponding to the signal named name in the recording specified by uuid.

If span is provided, this function returns the equivalent of load(dataset, uuid, name)[:, span], but potentially avoids loading the entire signal's worth of sample data if the underlying signal file format supports partial access/random seeks.

See also: deserialize_lpcm

load(dataset::Dataset, uuid::UUID, names[, span::AbstractTimeSpan])

Return Dict(name => load(dataset, uuid, name[, span]) for name in names).

load(dataset::Dataset, uuid::UUID[, span::AbstractTimeSpan])

Return load(dataset, uuid, names[, span]) where names is a list of all signal names in the recording specified by uuid.

Onda.store! — Function

store!(dataset::Dataset, uuid::UUID, name::Symbol, samples::Samples;
       overwrite::Bool=true)

Add name => samples.signal to dataset.recordings[uuid].signals and serialize samples.data to the proper file location within dataset.path.

If overwrite is false, an error is thrown if samples already exists in recording/dataset. Otherwise, existing entries matching samples.signal will be deleted and replaced with samples.

Base.delete! — Function

delete!(dataset::Dataset, uuid::UUID)

Delete the recording whose UUID matches uuid from dataset. This function removes the matching Recording object from dataset.recordings, as well as deletes the corresponding subdirectory in the dataset's recordings directory.

Onda.save_recordings_file — Function

save_recordings_file(dataset::Dataset)

Overwrite joinpath(dataset.path, "recordings.msgpack.zst") with the contents of dataset.recordings.

Onda Metadata Objects

Onda.Signal — Type

Signal

A type representing an individual Onda signal object. Instances contain the following fields, following the Onda specification for signal objects:

channel_names::Vector{Symbol}

sample_unit::Symbol

sample_resolution_in_unit::Float64

sample_type::DataType

sample_rate::UInt64

file_extension::Symbol

file_options::Union{Nothing,Dict{Symbol,Any}}

Onda.signal_from_template — Function

signal_from_template(signal::Signal;
                     channel_names=signal.channel_names,
                     sample_unit=signal.sample_unit,
                     sample_resolution_in_unit=signal.sample_resolution_in_unit,
                     sample_type=signal.sample_type,
                     sample_rate=signal.sample_rate,
                     file_extension=signal.file_extension,
                     file_options=signal.file_options)

Return a Signal where each field is mapped to the corresponding keyword argument.

Onda.Annotation — Type

Annotation <: AbstractTimeSpan

A type representing an individual Onda annotation object. Instances contain the following fields, following the Onda specification for annotation objects:

key::String

value::String

start_nanosecond::Nanosecond

stop_nanosecond::Nanosecond

Onda.Recording — Type

Recording{C}

A type representing an individual Onda recording object. Instances contain the following fields, following the Onda specification for recording objects:

duration_in_nanoseconds::Nanosecond

signals::Dict{Symbol,Signal}

annotations::Set{Annotation}

custom::C

Onda.annotate! — Function

annotate!(recording::Recording, annotation::Annotation)

Returns push!(recording.annotations, annotation).

Samples

Onda.Samples — Type

Samples(signal::Signal, encoded::Bool, data::AbstractMatrix)

Return a Samples instance with the following fields:

signal::Signal: The Signal object that describes the Samples instance.

encoded::Bool: If true, the values in data are LPCM-encoded as prescribed by the Samples instance's signal. If false, the values in data have been decoded into the signal's canonical units.

data::AbstractMatrix: A matrix of sample data. The i th row of the matrix corresponds to the ith channel in signal.channel_names, while the jth column corresponds to the jth multichannel sample.

Note that getindex and view are defined on Samples to accept normal integer indices, but also accept channel names for row indices and TimeSpan values for column indices:

julia> eeg
Samples (00:43:11.062500000):
  signal.channel_names: [:fp1, :f3, :c3, :p3, :f7, :t3, :t5, :o1, :fz, :cz,
                         :pz, :fp2, :f4, :c4, :p4, :f8, :t4, :t6, :o2]
  signal.sample_unit: :microvolt
  signal.sample_resolution_in_unit: 0.25
  signal.sample_type: Int16
  signal.sample_rate: 128 Hz
  signal.file_extension: :zst
  signal.file_options: nothing
  encoded: true
  data:
19×331656 Array{Int16,2}:
 -421  -416    51  -229  …   -164   -318   -644
 -866  -860  -401  -684     -1665  -1805  -2139
 -809  -776  -320  -641     -1402  -1571  -1892
 -698  -642  -191  -522     -1391  -1585  -1891
 -801  -778  -307  -622     -1275  -1452  -1771
 -340  -297   168  -160  …  -1012  -1202  -1514
 -594  -544   -86  -410      -802   -982  -1298
 -254  -180   270   -68        68   -141   -435
 -567  -547   -86  -396     -1439  -1602  -1924
 -620  -584  -129  -450     -1486  -1664  -1970
 -595  -807  -187  -341  …  -1672  -1441  -2113
 -407  -386    71  -216      -993  -1146  -1466
 -490  -475   -17  -310      -861  -1011  -1332
 -540  -555   -79  -359     -1077   -973  -1508
 -437  -394    60  -263     -1051  -1232  -1541
 -721  -692  -239  -536  …  -1681  -1838  -2151
 -646  -597  -145  -460     -1119  -1298  -1602
 -468  -411    32  -295      -224   -411   -695
 -855  -792  -354  -686     -1584  -1782  -2065

julia> eeg[[:fp1, 2, :fz, :o1, :cz], TimeSpan(Minute(20), duration(eeg))]
Samples (00:23:11.062500000):
  signal.channel_names: [:fp1, :f3, :fz, :o1, :cz]
  signal.sample_unit: :microvolt
  signal.sample_resolution_in_unit: 0.25
  signal.sample_type: Int16
  signal.sample_rate: 128 Hz
  signal.file_extension: :zst
  signal.file_options: nothing
  encoded: true
  data:
5×178056 Array{Int16,2}:
 -326    69   154  -281   …   -164   -318   -644
 -596  -204  -109  -543      -1665  -1805  -2139
 -472   -59    19  -422      -1439  -1602  -1924
 -453   -17    31  -418         68   -141   -435
 -797  -373  -306  -748      -1486  -1664  -1970

See also: encode, encode!, decode, decode!

Onda.channel — Function

channel(signal::Signal, name::Symbol)

Return i where signal.channel_names[i] == name.

channel(signal::Signal, i::Integer)

Return signal.channel_names[i].

channel(samples::Samples, name::Symbol)

Return channel(samples.signal, name).

This function is useful for indexing rows of samples.data by channel names.

channel(samples::Samples, i::Integer)

Return channel(samples.signal, i).

Onda.channel_count — Function

channel_count(signal::Signal)

Return length(signal.channel_names).

channel_count(samples::Samples)

Return channel_count(samples.signal).

Onda.sample_count — Function

sample_count(samples::Samples)

Return the number of multichannel samples in samples (i.e. size(samples.data, 2))

Onda.encode — Function

encode(sample_type::DataType, sample_resolution_in_unit, samples, dither_storage=nothing)

Return a copy of samples quantized according to sample_type and sample_resolution_in_unit. sample_type must be a concrete subtype of Onda.VALID_SAMPLE_TYPE_UNION. Quantization of an individual sample s is performed via:

round(S, s / sample_resolution_in_unit)

with additional special casing to clip values exceeding the encoding's dynamic range.

If dither_storage isa Nothing, no dithering is applied before quantization.

If dither_storage isa Missing, dither storage is allocated automatically and triangular dithering is applied to the signal prior to quantization.

Otherwise, dither_storage must be a container of similar shape and type to samples. This container is then used to store the random noise needed for the triangular dithering process, which is applied to the signal prior to quantization.

encode(samples::Samples, dither_storage=nothing)

If samples.encoded is false, return a Samples instance that wraps:

encode(samples.signal.sample_type,
       samples.signal.sample_resolution_in_unit,
       samples.data, dither_storage)

If samples.encoded is true, this function is the identity.

Onda.encode! — Function

encode!(result_storage, sample_type::DataType, sample_resolution_in_unit, samples, dither_storage=nothing)
encode!(result_storage, sample_resolution_in_unit, samples, dither_storage=nothing)

Similar to encode(sample_type, sample_resolution_in_unit, samples, dither_storage), but write encoded values to result_storage rather than allocating new storage.

sample_type defaults to eltype(result_storage) if it is not provided.

encode!(result_storage, samples::Samples, dither_storage=nothing)

If samples.encoded is false, return a Samples instance that wraps:

encode!(result_storage,
        samples.signal.sample_type,
        samples.signal.sample_resolution_in_unit,
        samples.data, dither_storage)`.

If samples.encoded is true, return a Samples instance that wraps copyto!(result_storage, samples.data).

Onda.decode — Function

decode(sample_resolution_in_unit, samples)

Return sample_resolution_in_unit .* samples

decode(samples::Samples)

If samples.encoded is true, return a Samples instance that wraps decode(samples.signal.sample_resolution_in_unit, samples.data).

If samples.encoded is false, this function is the identity.

Onda.decode! — Function

decode!(result_storage, sample_resolution_in_unit, samples)

Similar to decode(sample_resolution_in_unit, samples), but write decoded values to result_storage rather than allocating new storage.

decode!(result_storage, samples::Samples)

If samples.encoded is true, return a Samples instance that wraps decode!(result_storage, samples.signal.sample_resolution_in_unit, samples.data).

If samples.encoded is false, return a Samples instance that wraps copyto!(result_storage, samples.data).

AbstractTimeSpan

Onda.AbstractTimeSpan — Type

AbstractTimeSpan

A type repesenting a continuous, inclusive span between two points in time.

All subtypes of AbstractTimeSpan must implement:

first(::AbstractTimeSpan)::Nanosecond: return the first nanosecond contained in span

last(::AbstractTimeSpan)::Nanosecond: return the last nanosecond contained in span

For convenience, many Onda functions that accept AbstractTimeSpan values also accept Dates.Period values.

See also: TimeSpan

Onda.TimeSpan — Type

TimeSpan(first, last)

Return TimeSpan(Nanosecond(first), Nanosecond(last))::AbstractTimeSpan.

See also: AbstractTimeSpan

Onda.contains — Function

contains(a, b)

Return true if the timespan b lies entirely within the timespan a, return false otherwise.

Onda.overlaps — Function

overlaps(a, b)

Return true if the timespan a and the timespan b overlap, return false otherwise.

Onda.shortest_timespan_containing — Function

shortest_timespan_containing(spans)

Return the shortest possible TimeSpan containing all timespans in spans.

spans is assumed to be an iterable of timespans.

Onda.duration — Function

duration(span)

Return the duration of span as a Period.

For span::AbstractTimeSpan, this is equivalent to last(span) - first(span).

For span::Period, this function is the identity.

duration(recording::Recording)

Returns recording.duration_in_nanoseconds.

duration(samples::Samples)

Returns the Nanosecond value for which samples[TimeSpan(0, duration(samples))] == samples.data.

Serialization

Onda.AbstractLPCMSerializer — Type

AbstractLPCMSerializer

A type whose subtypes support:

deserialize_lpcm

serialize_lpcm

All definitions of subtypes of the form S<:AbstractLPCMSerializer must also support a constructor of the form S(::Signal).

Authors should additionally call register_file_extension_for_serializer to register their new serializer with Onda's internal global file extension -> serializer type map.

See also: serializer, LPCM, LPCMZst

Onda.serializer — Function

serializer(signal::Signal)

Return S(signal) where S is the AbstractLPCMSerializer that corresponds to signal.file_extension (as determined by the serializer author via register_file_extension_for_serializer).

See also: deserialize_lpcm, serialize_lpcm

Onda.deserialize_lpcm — Function

deserialize_lpcm(bytes, serializer::AbstractLPCMSerializer)

Return a channels-by-timesteps AbstractMatrix of interleaved LPCM-encoded sample data by deserializing the provided bytes from the given serializer.

Note that this operation may be performed in a zero-copy manner such that the returned sample matrix directly aliases bytes.

This function is the inverse of the corresponding serialize_lpcm method, i.e.:

serialize_lpcm(deserialize_lpcm(bytes, serializer), serializer) == bytes

deserialize_lpcm(bytes, serializer::AbstractLPCMSerializer, sample_offset, sample_count)

Similar to deserialize_lpcm(bytes, serializer), but deserialize only the segment requested via sample_offset and sample_count.

deserialize_lpcm(io::IO, serializer::AbstractLPCMSerializer[, sample_offset, sample_count])

Similar to the corresponding deserialize_lpcm(bytes, ...) methods, but the bytes to be deserialized are read directly from io.

If sample_offset/sample_count is provided and io/serializer support seeking, implementations of this method may read only the bytes required to extract the requested segment instead of reading the entire stream.

Onda.serialize_lpcm — Function

serialize_lpcm(samples::AbstractMatrix, serializer::AbstractLPCMSerializer)

Return the AbstractVector{UInt8} of bytes that results from serializing samples to the given serializer, where samples is a channels-by-timesteps matrix of interleaved LPCM-encoded sample data.

Note that this operation may be performed in a zero-copy manner such that the returned AbstractVector{UInt8} directly aliases samples.

This function is the inverse of the corresponding deserialize_lpcm method, i.e.:

deserialize_lpcm(serialize_lpcm(samples, serializer), serializer) == samples

serialize_lpcm(io::IO, samples::AbstractMatrix, serializer::AbstractLPCMSerializer)

Similar to the corresponding serialize_lpcm(samples, serializer) method, but serializes directly to io.

Onda.LPCM — Type

LPCM{S}(channel_count)
LPCM(signal::Signal)

Return a LPCM<:AbstractLPCMSerializer instance corresponding to Onda's default interleaved LPCM format assumed for signal files with the ".lpcm" extension.

S corresponds to signal.sample_type, while channel_count corresponds to signal.channel_names.

Note that bytes (de)serialized via this serializer are little-endian per the Onda specification.

Onda.LPCMZst — Type

LPCMZst(lpcm::LPCM; level=3)
LPCMZst(signal::Signal)

Return a LPCMZst<:AbstractLPCMSerializer instance that corresponds to Onda's default interleaved LPCM format compressed by zstd. This serializer is assumed for signal files with the ".lpcm.zst" extension.

The level keyword argument sets the same compression level parameter as the corresponding flag documented by the zstd command line utility.

See https://facebook.github.io/zstd/ for details about zstd.