API Documentation

Below is the documentation for all functions exported by Onda.jl. For general information regarding the Onda format, please see beacon-biosignals/OndaFormat.

Note that Onda.jl's API follows a specific philosophy with respect to property access: users are generally expected to access fields via Julia's object.fieldname syntax, but should only mutate objects via the exposed API methods documented below.

Dataset

Onda.Dataset — Type

Dataset(path; create=false)

Return a Dataset instance that contains all metadata necessary to read and write to the Onda dataset stored at path. Note that this constuctor loads all the Recording objects contained in path/recordings.msgpack.zst.

If create is true, then an empty Onda dataset will be created at path.

Onda.samples_path — Function

samples_path(dataset::Dataset, uuid::UUID)

Return the samples subdirectory path corresponding to the recording specified by uuid.

samples_path(dataset::Dataset, uuid::UUID, name::Symbol,
             file_extension=dataset.recordings[uuid].signals[name].file_extension)

Return the samples file path corresponding to the signal named name within the recording specified by uuid.

Onda.create_recording! — Function

create_recording!(dataset::Dataset, uuid::UUID=uuid4())

Create uuid::UUID => recording::Recording, add the pair to dataset.recordings, and return the pair.

Onda.load — Function

load(dataset::Dataset, uuid::UUID, name::Symbol[, span::AbstractTimeSpan])

Load and return the Samples object corresponding to the signal named name in the recording specified by uuid.

If span is provided, this function returns the equivalent of load(dataset, uuid, name)[:, span], but potentially avoids loading the entire signal's worth of sample data if the underlying signal file format supports partial access/random seeks.

See also: deserialize_lpcm

load(dataset::Dataset, uuid::UUID, names[, span::AbstractTimeSpan])

Return Dict(name => load(dataset, uuid, name[, span]) for name in names).

load(dataset::Dataset, uuid::UUID[, span::AbstractTimeSpan])

Return load(dataset, uuid, names[, span]) where names is a list of all signal names in the recording specified by uuid.

Onda.store! — Function

store!(dataset::Dataset, uuid::UUID, name::Symbol, samples::Samples;
       overwrite::Bool=true)

Add name => samples.signal to dataset.recordings[uuid].signals and serialize samples.data to the proper file location within dataset.path.

If overwrite is false, an error is thrown if samples already exists in recording/dataset. Otherwise, existing entries matching samples.signal will be deleted and replaced with samples.

Base.delete! — Function

delete!(dataset::Dataset, uuid::UUID)

Delete the recording whose UUID matches uuid from dataset. This function removes the matching Recording object from dataset.recordings, as well as deletes the corresponding subdirectory in the dataset's samples directory.

delete!(dataset::Dataset, uuid::UUID, name::Symbol)

Delete the signal whose name matches name from the recording whose UUID matches uuid in dataset. This function removes the matching Signal object from dataset.recordings[uuid], as well as deletes the corresponding sample data in the dataset's samples directory.

Onda.save_recordings_file — Function

save_recordings_file(dataset::Dataset)

Overwrite joinpath(dataset.path, "recordings.msgpack.zst") with the contents of dataset.recordings.

Onda Metadata Objects

Onda.Signal — Type

Signal

A type representing an individual Onda signal object. Instances contain the following fields, following the Onda specification for signal objects:

channel_names::Vector{Symbol}

start_nanosecond::Nanosecond

stop_nanosecond::Nanosecond

sample_unit::Symbol

sample_resolution_in_unit::Float64

sample_offset_in_unit::Float64

sample_type::DataType

sample_rate::Float64

file_extension::Symbol

file_options::Union{Nothing,Dict{Symbol,Any}}

Onda.signal_from_template — Function

signal_from_template(signal::Signal;
                     channel_names=signal.channel_names,
                     start_nanosecond=signal.start_nanosecond,
                     stop_nanosecond=signal.stop_nanosecond,
                     sample_unit=signal.sample_unit,
                     sample_resolution_in_unit=signal.sample_resolution_in_unit,
                     sample_offset_in_unit=signal.sample_offset_in_unit,
                     sample_type=signal.sample_type,
                     sample_rate=signal.sample_rate,
                     file_extension=signal.file_extension,
                     file_options=signal.file_options)

Return a Signal where each field is mapped to the corresponding keyword argument.

Onda.span — Function

span(signal::Signal)

Return TimeSpan(signal.start_nanosecond, signal.stop_nanosecond).

Onda.sizeof_samples — Function

sizeof_samples(signal::Signal)

Returns the expected size (in bytes) of the encoded Samples object corresponding to the entirety of signal:

sample_count(signal) * channel_count(signal) * sizeof(signal.sample_type)

Onda.Annotation — Type

Annotation <: AbstractTimeSpan

A type representing an individual Onda annotation object. Instances contain the following fields, following the Onda specification for annotation objects:

value::String

start_nanosecond::Nanosecond

stop_nanosecond::Nanosecond

Onda.Recording — Type

Recording

A type representing an individual Onda recording object. Instances contain the following fields, following the Onda specification for recording objects:

signals::Dict{Symbol,Signal}

annotations::Set{Annotation}

Onda.set_span! — Function

set_span!(recording::Recording, name::Symbol, span::AbstractTimeSpan)

Replace recording.signals[name] with a copy that has the start_nanosecond and start_nanosecond fields set to match the provided span. Returns the newly constructed Signal instance.

set_span!(recording::Recording, span::TimeSpan)

Return Dict(name => set_span!(recording, name, span) for name in keys(recording.signals))

Onda.annotate! — Function

annotate!(recording::Recording, annotation::Annotation)

Returns push!(recording.annotations, annotation).

Samples

Onda.Samples — Type

Samples(signal::Signal, encoded::Bool, data::AbstractMatrix)

Return a Samples instance with the following fields:

signal::Signal: The Signal object that describes the Samples instance.

encoded::Bool: If true, the values in data are LPCM-encoded as prescribed by the Samples instance's signal. If false, the values in data have been decoded into the signal's canonical units.

data::AbstractMatrix: A matrix of sample data. The i th row of the matrix corresponds to the ith channel in signal.channel_names, while the jth column corresponds to the jth multichannel sample.

Note that getindex and view are defined on Samples to accept normal integer indices, but also accept channel names for row indices and TimeSpan values for column indices; see Onda/examples/tour.jl for a comprehensive set of indexing examples.

See also: encode, encode!, decode, decode!

Onda.channel — Function

channel(signal::Signal, name::Symbol)

Return i where signal.channel_names[i] == name.

channel(signal::Signal, i::Integer)

Return signal.channel_names[i].

channel(samples::Samples, name::Symbol)

Return channel(samples.signal, name).

This function is useful for indexing rows of samples.data by channel names.

channel(samples::Samples, i::Integer)

Return channel(samples.signal, i).

Onda.channel_count — Function

channel_count(signal::Signal)

Return length(signal.channel_names).

channel_count(samples::Samples)

Return channel_count(samples.signal).

Onda.sample_count — Function

sample_count(signal::Signal)

Return the number of multichannel samples that fit within duration(signal) given signal.sample_rate.

sample_count(samples::Samples)

Return the number of multichannel samples in samples (i.e. size(samples.data, 2))

Warning

sample_count(samples) is not generally equivalent to sample_count(samples.signal); the former is the sample count of the entire original signal in the context of its parent recording, whereas the latter is actual number of multichannel samples in samples.data.

Onda.encode — Function

encode(sample_type::DataType, sample_resolution_in_unit, sample_offset_in_unit,
       samples, dither_storage=nothing)

Return a copy of samples quantized according to sample_type, sample_resolution_in_unit, and sample_offset_in_unit. sample_type must be a concrete subtype of Onda.VALID_SAMPLE_TYPE_UNION. Quantization of an individual sample s is performed via:

round(S, (s - sample_offset_in_unit) / sample_resolution_in_unit)

with additional special casing to clip values exceeding the encoding's dynamic range.

If dither_storage isa Nothing, no dithering is applied before quantization.

If dither_storage isa Missing, dither storage is allocated automatically and triangular dithering is applied to the signal prior to quantization.

Otherwise, dither_storage must be a container of similar shape and type to samples. This container is then used to store the random noise needed for the triangular dithering process, which is applied to the signal prior to quantization.

encode(samples::Samples, dither_storage=nothing)

If samples.encoded is false, return a Samples instance that wraps:

encode(samples.signal.sample_type,
       samples.signal.sample_resolution_in_unit,
       samples.signal.sample_offset_in_unit,
       samples.data, dither_storage)

If samples.encoded is true, this function is the identity.

Onda.encode! — Function

encode!(result_storage, sample_type::DataType, sample_resolution_in_unit,
        sample_offset_in_unit, samples, dither_storage=nothing)
encode!(result_storage, sample_resolution_in_unit, sample_offset_in_unit,
        samples, dither_storage=nothing)

Similar to encode(sample_type, sample_resolution_in_unit, sample_offset_in_unit, samples, dither_storage), but write encoded values to result_storage rather than allocating new storage.

sample_type defaults to eltype(result_storage) if it is not provided.

encode!(result_storage, samples::Samples, dither_storage=nothing)

If samples.encoded is false, return a Samples instance that wraps:

encode!(result_storage,
        samples.signal.sample_type,
        samples.signal.sample_resolution_in_unit,
        samples.signal.sample_offset_in_unit,
        samples.data, dither_storage)`.

If samples.encoded is true, return a Samples instance that wraps copyto!(result_storage, samples.data).

Onda.decode — Function

decode(sample_resolution_in_unit, sample_offset_in_unit, samples)

Return sample_resolution_in_unit .* samples .+ sample_offset_in_unit

decode(samples::Samples)

If samples.encoded is true, return a Samples instance that wraps decode(samples.signal.sample_resolution_in_unit, samples.signal.sample_offset_in_unit, samples.data).

If samples.encoded is false, this function is the identity.

Onda.decode! — Function

decode!(result_storage, sample_resolution_in_unit, sample_offset_in_unit, samples)

Similar to decode(sample_resolution_in_unit, sample_offset_in_unit, samples), but write decoded values to result_storage rather than allocating new storage.

decode!(result_storage, samples::Samples)

If samples.encoded is true, return a Samples instance that wraps decode!(result_storage, samples.signal.sample_resolution_in_unit, samples.signal.sample_offset_in_unit, samples.data).

If samples.encoded is false, return a Samples instance that wraps copyto!(result_storage, samples.data).

AbstractTimeSpan

Onda.AbstractTimeSpan — Type

AbstractTimeSpan

A type repesenting a continuous, inclusive span between two points in time.

All subtypes of AbstractTimeSpan must implement:

first(::AbstractTimeSpan)::Nanosecond: return the first nanosecond contained in span

last(::AbstractTimeSpan)::Nanosecond: return the last nanosecond contained in span

For convenience, many Onda functions that accept AbstractTimeSpan values also accept Dates.Period values.

See also: TimeSpan

Onda.TimeSpan — Type

TimeSpan(first, last)

Return TimeSpan(Nanosecond(first), Nanosecond(last))::AbstractTimeSpan.

See also: AbstractTimeSpan

Onda.contains — Function

contains(a, b)

Return true if the timespan b lies entirely within the timespan a, return false otherwise.

Onda.overlaps — Function

overlaps(a, b)

Return true if the timespan a and the timespan b overlap, return false otherwise.

Onda.shortest_timespan_containing — Function

shortest_timespan_containing(spans)

Return the shortest possible TimeSpan containing all timespans in spans.

spans is assumed to be an iterable of timespans.

Onda.duration — Function

duration(span)

Return the duration of span as a Period.

For span::AbstractTimeSpan, this is equivalent to last(span) - first(span).

For span::Period, this function is the identity.

duration(signal::Signal)

Return duration(span(signal)).

duration(recording::Recording)

Returns maximum(s -> s.stop_nanosecond, values(recording.signals)); throws an ArgumentError if recording.signals is empty.

duration(samples::Samples)

Returns the Nanosecond value for which samples[TimeSpan(0, duration(samples))] == samples.data.

Warning

duration(samples) is not generally equivalent to duration(samples.signal); the former is the duration of the entire original signal in the context of its parent recording, whereas the latter is the actual duration of samples.data given samples.signal.sample_rate.

Onda.time_from_index — Function

time_from_index(sample_rate, sample_index)

Given sample_rate in Hz and assuming sample_index > 0, return the earliest Nanosecond containing sample_index.

Examples:

julia> time_from_index(1, 1)
0 nanoseconds

julia> time_from_index(1, 2)
1000000000 nanoseconds

julia> time_from_index(100, 100)
990000000 nanoseconds

julia> time_from_index(100, 101)
1000000000 nanoseconds

time_from_index(sample_rate, sample_range::AbstractUnitRange)

Return the TimeSpan corresponding to sample_range given sample_rate in Hz:

julia> time_from_index(100, 1:100)
TimeSpan(0 nanoseconds, 1000000000 nanoseconds)

julia> time_from_index(100, 101:101)
TimeSpan(1000000000 nanoseconds, 1000000000 nanoseconds)

julia> time_from_index(100, 301:600)
TimeSpan(3000000000 nanoseconds, 6000000000 nanoseconds)

Onda.index_from_time — Function

index_from_time(sample_rate, sample_time)

Given sample_rate in Hz, return the integer index of the most recent sample taken at sample_time. Note that sample_time must be non-negative and support convert(Nanosecond, sample_time).

Examples:

julia> index_from_time(1, Second(0))
1

julia> index_from_time(1, Second(1))
2

julia> index_from_time(100, Millisecond(999))
100

julia> index_from_time(100, Millisecond(1000))
101

index_from_time(sample_rate, span::AbstractTimeSpan)

Return the UnitRange of indices corresponding to span given sample_rate in Hz:

julia> index_from_time(100, TimeSpan(Second(0), Second(1)))
1:100

julia> index_from_time(100, TimeSpan(Second(1), Second(1)))
101:101

julia> index_from_time(100, TimeSpan(Second(3), Second(6)))
301:600

Serialization

Onda.AbstractLPCMSerializer — Type

AbstractLPCMSerializer

A type whose subtypes support:

deserialize_lpcm

serialize_lpcm

All definitions of subtypes of the form S<:AbstractLPCMSerializer must also support a constructor of the form S(::Signal) and overload Onda.serializer_constructor_for_file_extension with the appropriate file extension.

See also: serializer, LPCM, LPCMZst

Onda.serializer_constructor_for_file_extension — Function

Onda.serializer_constructor_for_file_extension(::Val{:extension_symbol})

Return a constructor of the form S(::Signal)::AbstractLPCMSerializer corresponding to the provided extension.

This function should be overloaded for new AbstractLPCMSerializer subtypes.

Onda.serializer — Function

serializer(signal::Signal; kwargs...)

Return S(signal; kwargs...) where S is the AbstractLPCMSerializer that corresponds to signal.file_extension (as determined by the serializer author via serializer_constructor_for_file_extension).

See also: deserialize_lpcm, serialize_lpcm

Onda.deserialize_lpcm — Function

deserialize_lpcm(bytes, serializer::AbstractLPCMSerializer)

Return a channels-by-timesteps AbstractMatrix of interleaved LPCM-encoded sample data by deserializing the provided bytes from the given serializer.

Note that this operation may be performed in a zero-copy manner such that the returned sample matrix directly aliases bytes.

This function is the inverse of the corresponding serialize_lpcm method, i.e.:

serialize_lpcm(deserialize_lpcm(bytes, serializer), serializer) == bytes

deserialize_lpcm(bytes, serializer::AbstractLPCMSerializer, sample_offset, sample_count)

Similar to deserialize_lpcm(bytes, serializer), but deserialize only the segment requested via sample_offset and sample_count.

deserialize_lpcm(io::IO, serializer::AbstractLPCMSerializer[, sample_offset, sample_count])

Similar to the corresponding deserialize_lpcm(bytes, ...) methods, but the bytes to be deserialized are read directly from io.

If sample_offset/sample_count is provided and io/serializer support seeking, implementations of this method may read only the bytes required to extract the requested segment instead of reading the entire stream.

Onda.serialize_lpcm — Function

serialize_lpcm(samples::AbstractMatrix, serializer::AbstractLPCMSerializer)

Return the AbstractVector{UInt8} of bytes that results from serializing samples to the given serializer, where samples is a channels-by-timesteps matrix of interleaved LPCM-encoded sample data.

Note that this operation may be performed in a zero-copy manner such that the returned AbstractVector{UInt8} directly aliases samples.

This function is the inverse of the corresponding deserialize_lpcm method, i.e.:

deserialize_lpcm(serialize_lpcm(samples, serializer), serializer) == samples

serialize_lpcm(io::IO, samples::AbstractMatrix, serializer::AbstractLPCMSerializer)

Similar to the corresponding serialize_lpcm(samples, serializer) method, but serializes directly to io.

Onda.LPCM — Type

LPCM{S}(channel_count)
LPCM(signal::Signal)

Return a LPCM<:AbstractLPCMSerializer instance corresponding to Onda's default interleaved LPCM format assumed for signal files with the ".lpcm" extension.

S corresponds to signal.sample_type, while channel_count corresponds to signal.channel_names.

Note that bytes (de)serialized via this serializer are little-endian per the Onda specification.

Onda.LPCMZst — Type

LPCMZst(lpcm::LPCM; level=3)
LPCMZst(signal::Signal; level=3)

Return a LPCMZst<:AbstractLPCMSerializer instance that corresponds to Onda's default interleaved LPCM format compressed by zstd. This serializer is assumed for signal files with the ".lpcm.zst" extension.

The level keyword argument sets the same compression level parameter as the corresponding flag documented by the zstd command line utility.

See https://facebook.github.io/zstd/ for details about zstd.

Upgrading Older Datasets to Newer Datasets

Onda.upgrade_onda_format_from_v0_2_to_v0_3! — Function

Onda.upgrade_onda_format_from_v0_2_to_v0_3!(path, combine_annotation_key_value)

Upgrade the Onda v0.2 dataset at path to a Onda v0.3 dataset, returning the upgraded Dataset. This upgrade process overwrites path/recordings.msgpack.zst with a v0.3-compliant version of this file; for safety's sake, the old v0.2 file is preserved at path/old.recordings.msgpack.zst.backup.

A couple of the Onda v0.2 -> v0.3 changes require some special handling:

The custom field was removed from recording objects. This function thus writes out a file at path/recordings_custom.msgpack.zst that contains a map of UUIDs to corresponding recordings' custom values before deleting the custom field. This file can be deserialized via MsgPack.unpack(Onda.zstd_decompress(read("recordings_custom.msgpack.zst"))).

Annotations no longer have a key field. Thus, each annotation's existing key and value fields are combined into the single new value field via the provided callback combine_annotation_key_value(annotation_key, annotation_value).